+++ CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [5.0.2] - 2022-04-21
+
+### Fixed
+- Abstract blocks correct order of store registration.
+## [5.0.1] - 2022-04-20
+
+### Fixed
+- WP-CLI command for only necessary blocks and components in the init phase.
+
+## [5.0.0] - 2022-04-19
+
+- Major braking changes do to updates on css variables, and helpers and updating min PHP version to 7.4.
+- Full change log can be checked on Github [releases](https://github.com/infinum/eightshift-libs/releases/tag/5.0.0).
+
 ## [4.1.0] - 2022-02-03
 
 ### Added
@@ -304,6 +318,9 @@ Init setup
 
 [Unreleased]: https://github.com/infinum/eightshift-libs/compare/main...HEAD
 
+[5.0.2]: https://github.com/infinum/eightshift-libs/compare/5.0.1...v5.0.2
+[5.0.1]: https://github.com/infinum/eightshift-libs/compare/5.0.0...v5.0.1
+[5.0.0]: https://github.com/infinum/eightshift-libs/compare/4.1.0...v5.0.0
 [4.1.0]: https://github.com/infinum/eightshift-libs/compare/4.0.0...v4.1.0
 [4.0.0]: https://github.com/infinum/eightshift-libs/compare/3.1.0...v4.0.0
 [3.1.0]: https://github.com/infinum/eightshift-libs/compare/3.0.8...v3.1.0